chore: clean up middleware code
