+++ question-service/middlewares/logging.js
@@ -1,0 +1,7 @@
+const loggingMiddleware = (req, res, next) => {
+    console.log(`Received ${req.method} request to ${req.url}`);
+    next();
+  };
+  
+  export default loggingMiddleware;
+
